LayoutTests/storage/indexeddb/setversion-blocked-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test that setVersion is blocked on open connections.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "setversion-blocked.html"
ver = 1
blockedEventFired = false
indexedDB.open(dbname)

h1OpenOnSuccess():
h1 = event.target.result
indexedDB.open(dbname)

h2OpenOnSuccess():
h2 = event.target.result
h1.setVersion(String(ver++))

h2OnVersionChange():
old = 1
new = "1"

h1SetVersionOnBlocked():
blockedEventFired = true
h2.close()

h1SetVersionOnSuccess():

transactionOnComplete():
PASS blockedEventFired is true
PASS successfullyParsed is true

TEST COMPLETE
